Python/requirements.txt
# Python dependencies
google-cloud-texttospeech
openai
sounddevice
SpeechRecognition
pyaudio
serial
playsound
requests
